requirements.txt
Flask==2.1.3
pyotp==2.6.0
pillow==8.4.0
qrcode==7.3.1
python-dotenv==0.19.2
Werkzeug==2.3.8
